+++ code/src/Process.h
@@ -29,7 +29,7 @@ public:
     void set_current_machine_id(unsigned int current_machine_id);
     void set_process_move_cost(unsigned int pmc);
     void add_requirement(unsigned int requirement);
-    unsigned int get_requirement(unsigned int resource_id);
+    unsigned long int get_requirement(unsigned int resource_id);
 
     unsigned int get_id();
     unsigned int get_service_id();
